pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "predict-scaffold"
version = "0.1.0"
description = "PrediCT evaluation scaffold: preprocessing + data loading pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
]

[project.optional-dependencies]
medical = [
  "SimpleITK>=2.3",
  "pydicom>=2.4",
  "scipy>=1.10",
  "nibabel>=5.2",
  "opencv-python>=4.8",
  "tqdm>=4.65",
]
ml = [
  "torch>=2.1",
  "monai>=1.3",
  "scikit-learn>=1.3",
  "imbalanced-learn>=0.12",
  "pandas>=2.0",
]
coca_project1 = [
  "TotalSegmentator>=2.0.0",
  "matplotlib>=3.8",
  "pandas>=2.0",
]
coca_project2 = [
  "pyradiomics>=3.1",
  "scipy>=1.10",
  "matplotlib>=3.8",
  "pandas>=2.0",
  "seaborn>=0.13",
  "scikit-learn>=1.3",
  "umap-learn>=0.5",
]
coca_project3 = [
  "scipy>=1.10",
  "nibabel>=5.2",
  "pandas>=2.0",
  "matplotlib>=3.8",
]
coca = [
  "pandas>=2.0",
  "SimpleITK>=2.3",
  "opencv-python>=4.8",
  "tqdm>=4.65",
]
dev = [
  "ruff>=0.4",
]

[project.scripts]
predict = "predict.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
